+++ packages/core/button/src/types.ts
@@ -60,8 +60,7 @@ export interface ButtonProps {
   tabIndex?: number;
   /** Pass target down to a link within the button component, if a href is provided */
   target?: string;
-  /** Set whether it is a button or a form submission */
-  type?: 'button' | 'submit' | 'reset';
+
   /** Option to fit button width to its parent width */
   shouldFitContainer?: boolean;
   /** Pass in a custom theme */