+++ packages/elements/mention/examples/07-simple-mention.tsx
@@ -4,7 +4,7 @@ import { UIAnalyticsEventInterface } from '@atlaskit/analytics-next-types';
 import * as React from 'react';
 import { onMentionEvent } from '../example-helpers/index';
 import Mention from '../src/components/Mention';
-import { ELEMENTS_CHANNEL } from '../src/constants';
+import { ELEMENTS_CHANNEL } from '../src/_constants';
 import debug from '../src/util/logger';
 import { mockMentionData as mentionData } from '../src/__tests__/unit/_test-helpers';
 import { IntlProvider } from 'react-intl';