+++ packages/media/media-card/src/root/card/index.tsx
@@ -301,18 +301,17 @@ export class Card extends Component<CardProps, CardState> {
       analyticsEvent,
     };
 
-    const isVideo =
-      mediaItemDetails &&
-      (mediaItemDetails as FileDetails).mediaType === 'video';
-    // We want to block onClick because it is handled by inline video player
-    const shouldStartPlayingInline = useInlinePlayer && isVideo;
-    if (onClick && !shouldStartPlayingInline) {
+    if (onClick) {
       onClick(result, analyticsEvent);
     }
     if (!mediaItemDetails) {
       return;
     }
-    if (shouldStartPlayingInline) {
+
+    const isVideo =
+      mediaItemDetails &&
+      (mediaItemDetails as FileDetails).mediaType === 'video';
+    if (useInlinePlayer && isVideo) {
       this.setState({
         isPlayingFile: true,
       });